Refractor server
